04-py-api-docker-split-graphql/Borrow/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-stubs==5.1.0
django-stubs-ext==5.1.1
djangorestframework==3.15.2
djangorestframework-stubs==3.15.1
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy==1.11.2
mypy-extensions==1.0.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
sqlparse==0.5.2
types-PyYAML==6.0.12.20240917
types-requests==2.32.0.20241016
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
